feat(petra): verify corrected A9b sensitivity campaign (#137)
* feat(petra): add A9b sensitivity campaign harness

* fix(petra): dispatch A9b launcher by absolute path

* docs(program): record verified A9b negative result

+++ petra/crates/petra-deck/examples/corrosion_study.rs
@@ -287,7 +287,7 @@ fn run_replica(
         let name = names[fired.reaction as usize];
         if name == "metal_dissolution" {
             dissolution_events += 1;
-            dissolution_depth[fired.site as usize] += 1;
+            dissolution_depth[fired.site] += 1;
         } else if name.starts_with("repassivate_") {
             repassivation_events += 1;
         }
